added a message when offline money has been given

+++ src/main/java/fr/wakleg/market/event/LoginHandler.java
@@ -1,5 +1,6 @@
 package fr.wakleg.market.event;
 
+import fr.wakleg.ecobric.Main;
 import fr.wakleg.ecobric.util.IEntityDataSaver;
 import fr.wakleg.ecobric.util.MoneyManager;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerEntityEvents;
@@ -18,30 +19,32 @@ public class LoginHandler implements ServerEntityEvents.Load {
     public void onLoad(Entity entity, ServerWorld world) {
         if (entity instanceof ServerPlayerEntity){
             ServerPlayerEntity player = (ServerPlayerEntity) entity;
-            try {
-                Connection connection = DriverManager.getConnection(URL, USERNAME, PASSWORD);
+            if(!player.getWorld().isClient()){
+                try {
+                    Connection connection = DriverManager.getConnection(URL, USERNAME, PASSWORD);
 
-                String createTableQuery = "CREATE TABLE IF NOT EXISTS " + OFFLINE_MONEY_TABLE + " (player_uuid TEXT, amount INT)";
-                connection.createStatement().execute(createTableQuery);
+                    String createTableQuery = "CREATE TABLE IF NOT EXISTS " + OFFLINE_MONEY_TABLE + " (player_uuid TEXT, amount INT)";
+                    connection.createStatement().execute(createTableQuery);
 
-                String selectQuery = "SELECT amount FROM " + OFFLINE_MONEY_TABLE + " WHERE player_uuid = (?)";
-                PreparedStatement selectStatement = connection.prepareStatement(selectQuery);
-                selectStatement.setString(1, player.getUuidAsString());
-                ResultSet resultSet = selectStatement.executeQuery();
+                    String selectQuery = "SELECT amount FROM " + OFFLINE_MONEY_TABLE + " WHERE player_uuid = (?)";
+                    PreparedStatement selectStatement = connection.prepareStatement(selectQuery);
+                    selectStatement.setString(1, player.getUuidAsString());
+                    ResultSet resultSet = selectStatement.executeQuery();
 
-                if(resultSet.next()){
-                    MoneyManager.add((IEntityDataSaver) player, resultSet.getInt("amount"));
-                    String deleteQuery = "DELETE FROM " + OFFLINE_MONEY_TABLE + " WHERE (player_uuid) = (?)";
-                    PreparedStatement statement = connection.prepareStatement(deleteQuery);
-                    statement.setString(1, player.getUuidAsString());
-                    statement.executeUpdate();
+                    if(resultSet.next()){
+                        MoneyManager.add((IEntityDataSaver) player, resultSet.getInt("amount"));
+                        player.sendMessage(Text.literal("Vos items on été vendus pour un montant de " + resultSet.getInt("amount") + "$"));
 
-                    player.sendMessage(Text.literal("Vos items on été vendus pour un montant de " + resultSet.getInt("amount") + "$"));
+                        String deleteQuery = "DELETE FROM " + OFFLINE_MONEY_TABLE + " WHERE (player_uuid) = (?)";
+                        PreparedStatement statement = connection.prepareStatement(deleteQuery);
+                        statement.setString(1, player.getUuidAsString());
+                        statement.executeUpdate();
+                    }
+
+                    connection.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
                 }
-
-                connection.close();
-            } catch (SQLException e) {
-                e.printStackTrace();
             }
         }
     }
